Fixed file permissions.
There was a small bug in commit 7f8584758d72 where the 2nd chown/chmod was done on the uploadDir and not the imagePath as
before. So I made it to the chmod/chown on the imagePath again. But this had to be moved inside the "for member in
imageTarFile.getmembers() loop". Should be all better now.